display-drivers: protect setting fod hbm status by allow_bl_update
In the current implementation the fod hbm command might set the backlight
of the panel before receiving a frame from the driver during the first commit.
There is chance for garbage content to be shown onto the panel.
This change imposes a condition to delay the fod hbm command until the
first frame is received from the HW.

Change-Id: I5f1db2ac7c8a0c7cece1e773c87a92c81aee1561